Make SseStreamResource tests independent
The map of publishers in `SseStreamResource` is static and `publisherFor` will not register a new publisher for the same stream name. Before this fix, all tests were using the publisher registered by the first run test. The problem has not been evident until tearDown() phase was added with cleanup steps.

Signed-off-by: Jakub Zalas <jakub@zalas.pl>