Fix serving the directory index from a JAR
On some macos environments `/context` is not translated to `/context/index.html` due to a `NullPointerException` in the `java.io.FilterInputStream`:

```
18:38:08.200 [pool-2-thread-4] ERROR i.v.h.r.ResourceRequestHandlerActor - Runtime thrown by resource dispatcher
java.lang.NullPointerException: null
        at java.io.FilterInputStream.read(FilterInputStream.java:133)
        at java.io.FilterInputStream.read(FilterInputStream.java:107)
        at io.vlingo.http.resource.StaticFilesResource.isDirectory(StaticFilesResource.java:89)
        at io.vlingo.http.resource.StaticFilesResource.contentFilePath(StaticFilesResource.java:74)
        at io.vlingo.http.resource.StaticFilesResource.serveFile(StaticFilesResource.java:58)
        [...]
```

In the same time `/context/` (notice trailing `/`) is properly handled.

The problematic code in `isDirectory` checks if bytes can be read: `in.read(bytes) == -1`, but this does not work in the described scenario on macos.

Currently, code checks if the request path is a directory and appends `/index.html` if it is.

I suggested that we instead check if the request path is a file, otherwise append `index.html`.

Signed-off-by: Jakub Zalas <jakub@zalas.pl>